# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Parchat API.",
        "title": "Parchat API",
        "termsOfService": "https://pachat.online/terms/",
        "contact": {
            "name": "Parchat Support",
            "url": "https://pachat.online/support",
            "email": "parchat.soporte@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/chat/ws": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establece una conexión WebSocket para mensajería en tiempo real",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Conexión WebSocket para chat en tiempo real",
                "responses": {
                    "101": {
                        "description": "Switching Protocols a WebSocket",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}